AUTH SYSTEM:
Logging in authorizes the user, saves their image (to prevent Discord-images being changed).
THERE'S CURRENTLY NOT ANY DATA SAVING ADDED, SO YOU CAN ONLY LOG IN, IF YOU ADD THE USER MANUALLY!

Sessions-cookies are hashed and salted, and stored in the users browser.
Upon visiting the page, if a cookie is present in the browser, it will be sent to the server, checked for its validity.

If the cookie is valid, the user will be logged in, and the user's data will be fetched from the database.

STAFF:

Staff is fetched from the database, instead of being stored in a JSON file.

CALENDAR:

Minor changes to the calendar, to make it more user-friendly.

MODS:

Mods are fetched from the database, instead of being stored in a JSON file.
Added a script that auto-generates a .batch-file, that can be downloaded to check if the Plugins-folder contain any of the banned mods.

ADDED:

Website:
- Admin test-page, to test permissions.

Files:
- Database-schema for the database + DATABASE.MD to explain the arrays